requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0

langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
mcp>=1.0.0
langchain-mcp-adapters>=0.2.2
mcp>=1.2.0
pydantic>=2.7.0
langsmith

fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0